add: Are you sure you want to delete?

+++ src/History/History.js
@@ -1,10 +1,15 @@
 import React from "react";
 
 import Box from "@material-ui/core/Box";
+import Button from "@material-ui/core/Button";
 import IconButton from "@material-ui/core/IconButton";
 import ArrowRightAltIcon from "@material-ui/icons/ArrowRightAlt";
 import DeleteIcon from "@material-ui/icons/Delete";
 import EditIcon from "@material-ui/icons/Edit";
+import Dialog from "@material-ui/core/Dialog";
+import DialogActions from "@material-ui/core/DialogActions";
+import DialogContent from "@material-ui/core/DialogContent";
+import DialogContentText from "@material-ui/core/DialogContentText";
 import List from "@material-ui/core/List";
 import ListItem from "@material-ui/core/ListItem";
 import Grid from "@material-ui/core/Grid";
@@ -18,6 +23,9 @@ import TransferForm from "../Transfers/TransferForm";
 const useStyles = makeStyles({
   root: {
     width: 500,
+  },
+  modal: {
+    width: 420,
   },
 });
 
@@ -78,18 +86,67 @@ const History = (props) => {
                 </IconButton>
               </Grid>
               <Grid item xs={1} align="right">
-                <IconButton
-                  onClick={() =>
-                    props.deleteTransaction(
-                      transaction.id,
-                      transaction.Amount,
-                      transaction.From,
-                      transaction.To
-                    )
-                  }
-                >
+                <IconButton onClick={() => props.openModal(transaction)}>
                   <DeleteIcon />
                 </IconButton>
+                {props.transactionToDelete && (
+                  <Dialog open={props.showModal} onClose={props.closeModal}>
+                    <DialogContent>
+                      <DialogContentText>
+                        Are you sure you want to delete this transaction?
+                      </DialogContentText>
+                      <ListItem className={classes.modal}>
+                        <Grid item xs={2}>
+                          <Typography color="textSecondary" variant="body1">
+                            {`${props.transactionToDelete.Date.split(" ")[2]} ${
+                              props.transactionToDelete.Date.split(" ")[1]
+                            }`}
+                          </Typography>
+                        </Grid>
+                        <Grid item xs={2}>
+                          <Typography
+                            color={props.transactionColor}
+                            variant="body1"
+                          >{`${props.transactionToDelete.From}`}</Typography>
+                        </Grid>
+                        <Grid item xs={1}>
+                          <ArrowRightAltIcon color={props.arrowColor} />
+                        </Grid>
+                        <Grid item xs={2}>
+                          <Typography
+                            color={props.transactionColor}
+                            variant="body1"
+                          >{`${props.transactionToDelete.To}`}</Typography>
+                        </Grid>
+                        <Grid item xs={3}>
+                          <Typography
+                            color={props.amountColor}
+                            variant="body1"
+                            align="right"
+                          >{`${props.sign}${props.transactionToDelete.Amount} ILS`}</Typography>
+                        </Grid>
+                      </ListItem>
+                    </DialogContent>
+                    <DialogActions>
+                      <Button color="primary" onClick={props.closeModal}>
+                        Close
+                      </Button>
+                      <Button
+                        color="secondary"
+                        onClick={() =>
+                          props.deleteTransaction(
+                            props.transactionToDelete.id,
+                            props.transactionToDelete.Amount,
+                            props.transactionToDelete.From,
+                            props.transactionToDelete.To
+                          )
+                        }
+                      >
+                        Delete
+                      </Button>
+                    </DialogActions>
+                  </Dialog>
+                )}
               </Grid>
             </ListItem>
             <Typography color="textSecondary" variant="body2">
